arc.check_product {arcgisbinding}	R Documentation

ArcGIS product and license information

Description

Initialize connection to ArcGIS. Any script running directly from R (i.e. without being called from a Geoprocessing script) should first call arc.check_product to create a connection with ArcGIS. Provides installation details on the version of ArcGIS installed that arcgisbinding is communicating with.

Usage

arc.check_product()

Details

Returned details include:

Product: ArcGIS Desktop (i.e. ArcMap), or ArcGIS Pro. The name of the product connected.

License level: Basic, Standard, or Advanced are the three licensing levels available. Each provides progressively more functionality within the software. See the "Desktop Functionality Matrix" link for details.

Build number: The build number of the release being used. Useful in debugging and when creating error reports.

DLL: The dynamic linked library (DLL) in use allowing ArcGIS to communicate with R.

References

ArcGIS Desktop Functionality Matrix

Note

Additional license levels are available on ArcGIS Desktop: Server, EngineGeoDB, and Engine. These license levels are currently unsupported by this package.

Examples

info <- arc.check_product()
info$license # ArcGIS license level
info$version # ArcGIS build number
info$app # product name
info$dll # binding DLL in use

[Package arcgisbinding version 1.0.0.128 Index]