arc.dataset-class {arcgisbinding}	R Documentation

Class "arc.dataset"

Description

arc.dataset S4 class

Details

The dataset_type slot possible values are described in the referenced "dataset properties – data type" documentation. For feature datasets, extent contains four double values: (xmin, ymin, xmax, ymax). The fields slot includes the details of the ArcGIS data types of the relevant fields, which include data types not directly representable in R.

Slots

path

file path or layer name

dataset_type

dataset type

extent

spatial extent of the dataset

fields

list of field names

shapeinfo

geometry information (see arc.shapeinfo)

References

ArcGIS Help: Dataset properties – dataset type

Examples

ozone.file <- system.file("extdata", "ca_ozone_pts.shp",
                          package="arcgisbinding")
d <- arc.open(ozone.file)
names(d@fields) # get all field names
d@shapeinfo     # print shape info
d               # print dataset info

[Package arcgisbinding version 1.0.0.128 Index]