arc.progress_label {arcgisbinding}	R Documentation

Set progressor label for Geoprocessing dialog box

Description

Geoprocessing tools have a progressor, which includes both a progress label and a progress bar. The default progressor continuously moves back and forth to indicate the script is running. Using arc.progress_label and arc.progress_pos allows fine control over the script progress. Updating the progressor isn't necessary, but is useful in situations where solely outputting messages to the dialog is insufficient to communicate script progress.

Usage

arc.progress_label(label)

Arguments

label
Progress Label

Details

Using arc.progress_label allows control over the label that is displayed at the top of the running script. For example, it might be used to display the current step of the analysis taking place.

References

Understanding the progressor in script tools

Note

Currently only functions in ArcGIS Pro, and has no effect in ArcGIS Desktop.

This function is only available from within an ArcGIS session, and has no effect when run from the command line or in background geoprocessing.

See Also

arc.progress_pos, "Progress Messages" example Geoprocessing script

Examples

## Not run: 
arc.progress_label("Calculating bootstrap samples...")

## End(Not run)

[Package arcgisbinding version 1.0.0.128 Index]