arc.progress_pos {arcgisbinding}	R Documentation

Set progressor position for Geoprocessing dialog box

Description

Geoprocessing tools have a progressor, which includes both a progress label and a progress bar. The default progressor continuously moves back and forth to indicate the script is running. Using arc.progress_label and arc.progress_pos allow fine control over the script progress. Updating the progressor isn't necessary, but is useful in situations where solely outputting messages to the dialog is insufficient to communicate script progress.

Usage

arc.progress_pos(pos = -1)

Arguments

pos
Progress position (in percent)

Details

Using arc.progress_pos allows control over the progrssor position displayed at the top of the running script. The position is an integer percentage, 0 to 100, that the progress bar should be set to, with 100 indicating the script has completed (100%).

Setting the position to -1 resets the progressor to the default progressor, which continuously moves to indicate the script is running.

References

Understanding the progressor in script tools

Note

Currently only functions in ArcGIS Pro, and has no effect in ArcGIS Desktop.

This function is only available from within an ArcGIS session, and has no effect when run from the command line or in background geoprocessing.

See Also

arc.progress_label, "Progress Messages" example Geoprocessing script

Examples

## Not run: 
arc.progress_pos(55)

## End(Not run)

[Package arcgisbinding version 1.0.0.128 Index]