arc.shape-class {arcgisbinding}	R Documentation

Class "arc.shape"

Description

arc.shape is geometry collection

Note

arc.shape is attached to an ArcGIS data.frame as the attribute "shape". Each element corresponds to one record in the input data frame. Points are presented as an array of lists, with each list containing (x, y, Z, M), where

Examples

d <- arc.open(system.file("extdata", "ca_ozone_pts.shp",
                          package="arcgisbinding"))
arc.df <- arc.select(d, "FID")

shape <- arc.shape(arc.df)
# create data.frame with X and Y columns
df <- data.frame(arc.df, X=shape$x, Y=shape$y)
# print out row #42
df[42,]

[Package arcgisbinding version 1.0.0.128 Index]