arc.shapeinfo {arcgisbinding}	R Documentation

Shape Information

Description

arc.shapeinfo provides details on what type of geometry is stored within the dataset, and the spatial reference of the geometry. The well-known text, WKT, allows interoperable transfer of the spatial reference system (CRS) between environments. The WKID is a numeric value that ArcGIS uses to precisely specify a projection.

Usage

arc.shapeinfo(object)

Arguments

object
arc.dataset-class object

Slots

type

geometry type: "Point", "Polyline", or "Polygon"

hasZ

TRUE if geometry includes Z-values

hasM

TRUE if geometry includes M-values

WKT

well-known text representation of the shape's spatial reference

WKID

well-known ID of the shape's spatial reference

References

ArcGIS REST API: Using spatial references

Spatial reference lookup

See Also

arc.dataset-class arc.shape-class

Examples

d <- arc.open(system.file("extdata", "ca_ozone_pts.shp",
                          package="arcgisbinding"))
# from arc.dataset
info <- arc.shapeinfo(d)
info$WKT   # print dataset spatial reference

# from arc.shape
df <- arc.select(d, 'ozone')
arc.shapeinfo(arc.shape(df))

[Package arcgisbinding version 1.0.0.128 Index]