arcgisbinding {arcgisbinding}	R Documentation

Bindings for ArcGIS

Description

This package provides classes for loading, converting and exporting ArcGIS datasets and layers in R.

Introduction

For a complete list of exported functions, use library(help = "arcgisbinding").

References

sp package

CRAN Task View: Analysis of Spatial Data

[Package arcgisbinding version 1.0.0.128 Index]